[COSMETIC CHANGE ONLY] workaround for Eclipse compiler

+++ apps/experiments/3DViewer/src/main/java/com/javafx/experiments/jfx3dviewer/TimelineController.java
@@ -49,7 +49,7 @@ public class TimelineController {
     private final ChangeListener<Number> rateListener = new ChangeListener<Number>() {
         @Override public void changed(ObservableValue<? extends Number> observable, Number oldValue, Number newRate) {
             System.out.println("newRate = " + newRate);
-            if (newRate == 0 && playBtn.isSelected()) {
+            if (newRate.intValue() == 0 && playBtn.isSelected()) {
                 playBtn.setSelected(false);
             }
         }
